requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anticaptchaofficial==1.0.50
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.133
botocore==1.29.133
bs4==0.0.1
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
choicesenum==0.7.0
cryptocode==0.1
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-admin-multi-select-filter==1.0.0
django-admin-rangefilter==0.10.0
django-cors-headers==4.0.0
django-cron==0.6.0
django-crontab==0.7.1
django-csp==3.7
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
frozenlist==1.3.3
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-auth==2.18.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-translate==3.11.1
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
MarkupPy==1.14
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.24.3
odfpy==1.4.1
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.1
pdfkit==1.0.0
Pillow==10.0.1
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
PyJWT==2.7.0
pymemcache==4.0.0
PyMySQL==1.0.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
razorpay==1.4.1
redis==4.5.5
requests==2.30.0
retrying==1.3.4
rsa==4.9
s3transfer==0.6.1
selenium==4.9.1
sentry-sdk==1.22.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.4.0
trio==0.22.0
trio-websocket==0.10.2
twilio==8.2.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.15
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.2
